requirements.txt
boto3==1.26.71
fastapi==0.92.0
Jinja2==3.1.2
Pillow==9.4.0
python-dotenv==0.21.1
